fix(whatsapp): split media idle helper and clear proof oxlint
